### 2024-11-22 #7517 @editorialbot
[PRE REVIEW]: ConFluxPro: A toolkit for soil gas analysis
**Submitting author:** @valentingar (<a href="http://orcid.org/0000-0001-5320-374X">Valentin Gartiser</a>)
**Repository:** https://github.com/valentingar/confluxpro/
**Branch with paper.md** (empty if default branch): joss_submission
**Version:** 1.2.1
