+++ pkgdown.yml
@@ -7,7 +7,7 @@ articles:
   data_preparation: data_preparation.html
   flux: flux.html
   uncertainty: uncertainty.html
-last_built: 2024-11-12T11:27Z
+last_built: 2024-11-27T08:32Z
 urls:
   reference: https://valentingar.github.io/ConFluxPro/reference
   article: https://valentingar.github.io/ConFluxPro/articles
