### 2024-11-27 #7517 @valentingar
Hello @kthyng!
Thank you for pointing that out! I've updated the ORCIDs and the paper should now have compiled correctly.
