### 2024-11-27 #7517 @valentingar
**Potential Reviewers:**
- junbinzhao - Author of [FluxCalR](https://github.com/junbinzhao/FluxCalR) - [published in JOSS](https://joss.theoj.org/papers/10.21105/joss.01751)
- jmzobitz - Author of [neonSoilFlux](https://github.com/jmzobitz/neonSoilFlux)

- [Roland Fuß](https://www.thuenen.de/de/fachinstitute/agrarklimaschutz/personal/wissenschaftliches-personal/fuss-roland) - Author of [gasfluxes](https://git-dmz.thuenen.de/fuss/gasfluxes)

**And from the [list](https://reviewers.joss.theoj.org/lookup) searching for keywords (soil, gas, flux, CO2, CH4, methane):**
- Saryace
