### 2024-12-03 4a73369
Added missing DOI and ORCID, fixed formatting and capitalization of references
