+++ paper.bib
@@ -68,7 +68,7 @@
   pages = {8653}
 }
 @article{Jochheim2022,
-  title = {Dynamics of Soil $\mathrm{CO_2}$ Efflux and Vertical $\mathrm{CO_2}$ Production in a European Beech and a Scots Pine Forest},
+  title = {Dynamics of Soil $\mathrm{CO_2}$ Efflux and Vertical $\mathrm{CO_2}$ Production in a {European} Beech and a Scots Pine Forest},
   volume = {5},
   ISSN = {2624-893X},
   url = {http://dx.doi.org/10.3389/ffgc.2022.826298},
@@ -94,7 +94,7 @@
   pages = {169–183}
 }
 @article{Zhao2019,
-  title = {FluxCalR: a R package for calculating CO~2~ and CH~4~ fluxes from static chambers},
+  title = {FluxCalR: a {R} package for calculating $\mathrm{CO_2}$ and $\mathrm{CO_4}$ fluxes from static chambers},
   volume = {4},
   ISSN = {2475-9066},
   url = {http://dx.doi.org/10.21105/joss.01751},
@@ -133,7 +133,7 @@ yourself,  regardless of user experience},
   month = dec
 }
 @misc{Zobitz2024,
-  title = {neonSoilFlux: Compute Soil Carbon Fluxes for the National Ecological Observatory Network Sites},
+  title = {neonSoilFlux: Compute Soil Carbon Fluxes for the {National Ecological Observatory Network} Sites},
   url = {http://dx.doi.org/10.32614/CRAN.package.neonSoilFlux},
   DOI = {10.32614/cran.package.neonsoilflux},
   journal = {CRAN: Contributed Packages},
@@ -143,7 +143,8 @@ yourself,  regardless of user experience},
   month = may
 }
 @BOOK{Bittelli2015,
-  title     = "Soil physics with python",
+  title     = "Soil physics with {python}",
+  DOI       = "10.1093/acprof:oso/9780199683093.001.0001",
   author    = "Bittelli, Marco and Campbell, Gaylon S and Tomei, Fausto",
   publisher = "Oxford University Press",
   month     =  dec,
