### 2024-12-03 #7517 @valentingar
Hello @kthyng - great news! Thank you for considering our submission. We are looking forward to the review.

I fixed the @editorialbot comments and capitalization in the references and will regenerate the pdf now.
valentingar/ConFluxPro@4a73369
