### 2024-12-12 6538bbc
more efficient handling of many groups in pro_flux()
