### 2024-12-13 e14cd18
added logic to parallelize over groups in pro_flux() if there are many and one group does not stand out
