### 2025-03-10 #7517 @kanishkan91
@feldl are you interested in reviewing this paper for JOSS based on your background, previous publications in the journal?
