### 2025-03-26 #7517 @kanishkan91
@dankusanovic would you be interested in reviewing this paper and package for JOSS?
