### 2025-03-26 #7517 @kanishkan91
@naingeet  would you be interested in reviewing this paper and package for JOSS?
