### 2025-03-31 1821876
switched to number fraction way of calculating in prod_mod_conc
