### 2025-04-01 ab69308
changed pro_flux output to x_ppm and conc
