### 2025-04-02 #7517 @kanishkan91
@jnnsbrr would you be interested in reviewing this paper for JOSS given your recent publication?
