### 2025-04-02 #7517 @naingeet
Hi,

Sorry for delay in replay. I am working towards a publication myself so
work schedule is bit tight for next couple of weeks.

Regards,
Geeta

On Tue, Mar 25, 2025 at 11:24 PM Kanishka Narayan ***@***.***>
wrote:

> @naingeet <https://github.com/naingeet> would you be interested in
> reviewing this paper and package for JOSS?
>
