### 2025-04-23 #7517 @kanishkan91
@shubhamjain15 would you be interested in reviewing this paper for JOSS given your previous reviews with the journal? It seems like a good match to your profile.
