### 2025-04-23 #7517 @kanishkan91
@ciara-donegan Would you be interested in reviewing this paper for JOSS?
