### 2025-04-23 #7517 @kanishkan91
@patrickcgray would you be interested in reviewing this paper for JOSS?
