### 2025-04-23 #7517 @ciara-donegan
> [@ciara-donegan](https://github.com/ciara-donegan) Would you be interested in reviewing this paper for JOSS?

Sure, I'd be happy to review!
