### 2025-04-23 #8094 @editorialbot
[REVIEW]: ConFluxPro: A toolkit for soil gas analysis
**Submitting author:** @valentingar (<a href="http://orcid.org/0000-0001-5320-374X">Valentin Gartiser</a>)
**Repository:** https://github.com/valentingar/confluxpro/
**Branch with paper.md** (empty if default branch): joss_submission
**Version:** v1.2.2
