### 2025-05-16 #8094 @shubhamjain15
Hi @valentingar, please consider updating the installation instructions to - ``remotes::install_github("valentingar/ConFluxPro", build_vignettes = TRUE, dependencies = TRUE)``  to ensure the vignettes build smoothly for users, as the vignettes won’t build or load properly due to missing suggested packages used in them. Also, I am seeing ``vignette("overview", package = "ConFluxPro")`` in the instructions, but there is no vignette named "overview".

![Image](https://github.com/user-attachments/assets/47aaebb8-7877-4f6a-b6ee-2c0b74b868aa)
