### 2025-05-20 f153643
switched to x_ppm calculation method
