### 2025-05-20 ab33f32
updated installation instructions in readme
