### 2025-05-20 3ccc661
updated installation instructions in readme
