### 2025-05-20 #8094 @valentingar
> Hi [@valentingar](https://github.com/valentingar), please consider updating the installation instructions to - `remotes::install_github("valentingar/ConFluxPro", build_vignettes = TRUE, dependencies = TRUE)` to ensure the vignettes build smoothly for users, as the vignettes won’t build or load properly due to missing suggested packages used in them. Also, I am seeing `vignette("overview", package = "ConFluxPro")` in the instructions, but there is no vignette named "overview".
>
> ![Image](https://github.com/user-attachments/assets/47aaebb8-7877-4f6a-b6ee-2c0b74b868aa)

Hi @shubhamjain15,
thanks for the suggestion! I've updated the instructions in the branch that holds the paper. I'll collect all changes in the review process there and merge into the main branch when we are finished.

@ciara-donegan @shubhamjain15,
thank you for giving your time to review this submission! Feel free to open issues in the package repository for any problems or improvements that arise.
