### 2025-06-24 0f4c5f6
Solves `future` issue and closes #109
