### 2025-06-24 db0a166
Solves `future` issue and closes #109
