### 2025-06-24 9dc7128
Increment version number to 1.2.2
