+++ pkgdown.yml
@@ -1,5 +1,5 @@
 pandoc: 3.1.11
-pkgdown: 2.1.1
+pkgdown: 2.1.3
 pkgdown_sha: ~
 articles:
   calibration: calibration.html
@@ -7,7 +7,7 @@ articles:
   data_preparation: data_preparation.html
   flux: flux.html
   uncertainty: uncertainty.html
-last_built: 2024-11-27T08:32Z
+last_built: 2025-06-24T11:09Z
 urls:
   reference: https://valentingar.github.io/ConFluxPro/reference
   article: https://valentingar.github.io/ConFluxPro/articles
