+++ pkgdown.yml
@@ -7,7 +7,7 @@ articles:
   data_preparation: data_preparation.html
   flux: flux.html
   uncertainty: uncertainty.html
-last_built: 2025-06-24T11:09Z
+last_built: 2025-06-24T11:13Z
 urls:
   reference: https://valentingar.github.io/ConFluxPro/reference
   article: https://valentingar.github.io/ConFluxPro/articles
