### 2025-06-24 #8094 @kanishkan91
@valentingar  I am recommending this for acceptance now. I will be reading through the paper for other typos etc. shortly. The AEiC in this submission track will review shortly and if all goes well this will go live soon! Big thank you to @shubhamjain15  and @ciara-donegan for reviewing! JOSS is volunteer run and relies heavily on researchers such as yourself.
