### 2025-06-24 #8094 @editorialbot
:wave: @openjournals/ese-eics, this paper is ready to be accepted and published.
