### 2025-06-24 #8094 @kthyng
@valentingar Please check the capitalization in your references. You can preserve capitalization by placing {} around characters/words in your .bib file. For example "scots", and not sure if "python" should be capitalized. Please check thoroughly.
