### 2025-06-25 0c6d86c
Enforce capitalization
