+++ paper.bib
@@ -41,7 +41,7 @@
   pages = {944–956}
 }
 @article{Maier2014,
-  title = {Using the gradient method to determine soil gas flux: A review},
+  title = {Using the {Gradient Method} to determine soil gas flux: A review},
   volume = {192–193},
   ISSN = {0168-1923},
   url = {http://dx.doi.org/10.1016/j.agrformet.2014.03.006},
@@ -68,7 +68,7 @@
   pages = {8653}
 }
 @article{Jochheim2022,
-  title = {Dynamics of Soil $\mathrm{CO_2}$ Efflux and Vertical $\mathrm{CO_2}$ Production in a {European} Beech and a Scots Pine Forest},
+  title = {Dynamics of Soil $\mathrm{CO_2}$ Efflux and Vertical $\mathrm{CO_2}$ Production in a {European} {Beech} and a {Scots Pine} Forest},
   volume = {5},
   ISSN = {2624-893X},
   url = {http://dx.doi.org/10.3389/ffgc.2022.826298},
@@ -143,11 +143,24 @@ yourself,  regardless of user experience},
   month = may
 }
 @BOOK{Bittelli2015,
-  title     = "Soil physics with {python}",
+  title     = {Soil physics with {Python}},
   DOI       = "10.1093/acprof:oso/9780199683093.001.0001",
   author    = "Bittelli, Marco and Campbell, Gaylon S and Tomei, Fausto",
   publisher = "Oxford University Press",
   month     =  dec,
   year      =  2015,
   address   = "London, England"
+}
+@article{Gartiser2025,
+  title = {A New Approach to Address Soil Heterogeneity as a Source of Uncertainty in the {Flux‐Gradient Method}: $\mathrm{CO_2}$ Case Studies},
+  volume = {76},
+  ISSN = {1365-2389},
+  url = {http://dx.doi.org/10.1111/ejss.70126},
+  DOI = {10.1111/ejss.70126},
+  number = {3},
+  journal = {European Journal of Soil Science},
+  publisher = {Wiley},
+  author = {Gartiser,  Valentin and Lang,  Verena and Maier,  Martin},
+  year = {2025},
+  month = may
 }
