### 2025-06-25 #8094 @valentingar
@kthyng, I updated the paper to enforce capitalization for proper names (including Python and Gradient Method).
Note, that I also added a reference to the scientific paper we had submitted to the European Journal of Soil Science alongside this manuscript and mentioned at submission. The paper has been published in the meantime and I think it makes sense to reference here as well since it gives the scientific background for ConFluxPro. If it is your opinion that this should not be added, because the review is already finished, I will remove it in a new commit.
